+++ DelvUI/Interface/PartyCooldowns/PartyCooldownsConfig.cs
@@ -156,10 +156,25 @@ namespace DelvUI.Interface.PartyCooldowns
 
             foreach (uint key in DefaultCooldowns.Keys)
             {
-                if (Cooldowns.Any(data => data.ActionId == key)) { continue; }
+                PartyCooldownData? data = Cooldowns.FirstOrDefault(data => data.ActionId == key);
+                PartyCooldownData defaultData = DefaultCooldowns[key];
 
-                Cooldowns.Add(DefaultCooldowns[key]);
-                needsSave = true;
+                if (data == null)
+                {
+                    Cooldowns.Add(defaultData);
+                }
+                else if (data != null && !data.Equals(defaultData))
+                {
+                    data.RequiredLevel = defaultData.RequiredLevel;
+                    data.JobId = defaultData.JobId;
+                    data.JobIds = defaultData.JobIds;
+                    data.Role = defaultData.Role;
+                    data.Roles = defaultData.Roles;
+                    data.CooldownDuration = defaultData.CooldownDuration;
+                    data.EffectDuration = defaultData.EffectDuration;
+
+                    needsSave = true;
+                }
             }
 
             ExcelSheet<Action>? sheet = Plugin.DataManager.GetExcelSheet<Action>();
@@ -363,10 +378,10 @@ namespace DelvUI.Interface.PartyCooldowns
 
             // RANGED
             [118] = NewData(118, JobIDs.BRD, 50, 120, 15, 30, 3), // battle voice
-            [7405] = NewData(7405, JobIDs.BRD, 88, 90, 15, 70, 2), // troubadour
+            [7405] = NewData(7405, JobIDs.BRD, 62, 90, 15, 70, 2), // troubadour
             [7408] = NewData(7408, JobIDs.BRD, 66, 90, 15, 40, 2), // nature's minne
             [25785] = NewData(25785, JobIDs.BRD, 90, 110, 15, 30, 3), // radiant finale
-            [16012] = NewData(16012, JobIDs.DNC, 88, 90, 15, 70, 2), // shield samba
+            [16012] = NewData(16012, JobIDs.DNC, 56, 90, 15, 70, 2), // shield samba
             [16004] = NewData(16004, JobIDs.DNC, 70, 120, 20, 30, 3), // technical step / finish
             [16889] = NewData(16889, JobIDs.MCH, 56, 90, 15, 70, 2), // tactician